First attempt at padding variables
The variables to pad are determined from the region that the current cursor is in. There is a full implementation of the refactoring workflow. The biggest obstacle right now is that the #pragma omp directive are respected during a refactoring. This is the default behavior of the rewriter.

There are also a bunch of TODO's left in the files to be followed-up on.

This commit belongs to issue #2.

Conflicts:

	edu.illinois.parallelism.openmp.refactoring.ui/META-INF/MANIFEST.MF
	edu.illinois.parallelism.openmp.refactoring.ui/src/edu/illinois/parallelism/openmp/refactoring/padscalarvariables/Messages.java
	edu.illinois.parallelism.openmp.refactoring.ui/src/edu/illinois/parallelism/openmp/refactoring/padscalarvariables/PadScalarVariablesRefactoring.java
	edu.illinois.parallelism.openmp.refactoring.ui/src/edu/illinois/parallelism/openmp/refactoring/padscalarvariables/PadScalarVariablesRunner.java