Initial attemp at a reusable minimal-change generator
The changes that CDT creates are too coarse-grained and tend to override
more than necessary. The changes are based on creating nested AST
changes. This also has the consequence of wiping out any
preprocessor directives since those aren't part of the AST per-se.

Here I have attempted to create the minimal change necessary based
on the node that we are inserting.

This commit belongs to issue #3.