Properly resize the refactoring dialog
The key insight is to call getShell().layout(true, true) once everything has been laid out. However, we don't have a hook into that even so right now we are excessively calling it during an event when the user selects (checks/unchecks) a variable to pad.

We can revisit this issue in the future but it works for the time being.

This commit belongs to issue #2.