Enable user to hit back/next on the Refactoring Wizard
Initially, I was using only one CCompositeChange object and each time a refactoring was previewed, it added the change to this single object.  However, the Refactoring Wizard expects a fresh copy of a CCompositeChange each time it previews/finishes.

To fix this, I create a new CCompositeChange each time the createChange method is called.

This commit belongs to issue #4.